Scalar functions post-integration changes
This is a follow-up to PR #630.

It makes the following changes to newly added Scalar Functions Java
API:

 - moves exception and registered shell classes into the
   `DuckDBFunction.java`
 - removes abstract classes for vector reader and writer
 - removes `DuckDBScalarContext` and `DuckDBScalarRow` in favour of
   streaming plain indices (as a `LongStream`) from the input data
   chunk; the row object inteface appeared to have an unintended
   overhead of creating a Java object for every input row that we would
   like to avoid. And without it the context abstraction appeared to be
   unnecessary

Null-propagation handling is changed the following way:

 - null propagation on Java side is enabled only for primitive callbacks
   (set automatically) and not exposed to the user, null propagation
   support for object callbacks is removed
 - null propagation on DuckDB engine side (
 `duckdb_scalar_function_set_special_handling` C API call skip) is also
 enabled automatically only for primitive callbacks, but it is
 additionally exposed to users as `withNullInNullOut()` builder call
 (replaces awkwardly named `withSpecialHandling()`); in some cases NULLs
 still can be passed to callbacks wheh `withNullInNullOut()` is
 set so callback still must check for nulls

Testing: more tests added aroung the null handling
