+++ src/test/java/org/duckdb/TestBindings.java
@@ -9,9 +9,7 @@ import static org.duckdb.test.Assertions.*;
 import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
-import java.nio.charset.StandardCharsets;
 import java.sql.*;
-import java.util.Arrays;
 
 public class TestBindings {
 
@@ -27,57 +25,22 @@ public class TestBindings {
         ByteBuffer inputVec = duckdb_create_vector(lt);
         ByteBuffer outputVec = duckdb_create_vector(lt);
 
-        DuckDBWritableVector input = new DuckDBWritableVectorImpl(inputVec, 3);
+        DuckDBWritableVector input = new DuckDBWritableVector(inputVec, 3);
         input.setInt(0, 1);
         input.setInt(1, 41);
         input.setInt(2, -5);
 
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(inputVec, 3);
-        DuckDBWritableVector output = new DuckDBWritableVectorImpl(outputVec, 3);
-        readable.rowIndexStream().forEachOrdered(row -> { output.setInt(row, readable.getInt(row) + 1); });
+        DuckDBReadableVector readable = new DuckDBReadableVector(inputVec, 3);
+        DuckDBWritableVector output = new DuckDBWritableVector(outputVec, 3);
+        readable.stream().forEachOrdered(row -> { output.setInt(row, readable.getInt(row) + 1); });
 
-        DuckDBReadableVector result = new DuckDBReadableVectorImpl(outputVec, 3);
+        DuckDBReadableVector result = new DuckDBReadableVector(outputVec, 3);
         assertEquals(result.getInt(0), 2);
         assertEquals(result.getInt(1), 42);
         assertEquals(result.getInt(2), -4);
 
         duckdb_destroy_vector(outputVec);
         duckdb_destroy_vector(inputVec);
-        duckdb_destroy_logical_type(lt);
-    }
-
-    public static void test_bindings_writable_vector_append_after_indexed_write() throws Exception {
-        ByteBuffer lt = duckdb_create_logical_type(DUCKDB_TYPE_INTEGER.typeId);
-        ByteBuffer vec = duckdb_create_vector(lt);
-
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, 3);
-        writable.setNull(0);
-        writable.setInt(1, 41);
-        writable.addInt(42);
-
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(vec, 3);
-        assertTrue(readable.isNull(0));
-        assertEquals(readable.getInt(1), 41);
-        assertEquals(readable.getInt(2), 42);
-
-        duckdb_destroy_vector(vec);
-        duckdb_destroy_logical_type(lt);
-    }
-
-    public static void test_bindings_writable_vector_failed_append_does_not_advance() throws Exception {
-        ByteBuffer lt = duckdb_create_logical_type(DUCKDB_TYPE_INTEGER.typeId);
-        ByteBuffer vec = duckdb_create_vector(lt);
-
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, 2);
-        assertThrows(() -> { writable.addString("boom"); }, DuckDBFunctionException.class);
-        writable.addInt(7);
-        writable.addInt(8);
-
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(vec, 2);
-        assertEquals(readable.getInt(0), 7);
-        assertEquals(readable.getInt(1), 8);
-
-        duckdb_destroy_vector(vec);
         duckdb_destroy_logical_type(lt);
     }
 
@@ -194,11 +157,11 @@ public class TestBindings {
         ByteBuffer vec = duckdb_create_vector(lt);
 
         long rowCount = duckdb_vector_size();
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, rowCount);
+        DuckDBWritableVector writable = new DuckDBWritableVector(vec, rowCount);
         writable.setNull(0);
         writable.setString(1, "duckdb");
 
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(vec, rowCount);
+        DuckDBReadableVector readable = new DuckDBReadableVector(vec, rowCount);
         assertNull(readable.getString(0));
         assertEquals(readable.getString(1), "duckdb");
         assertThrows(() -> { readable.getString(rowCount); }, IndexOutOfBoundsException.class);
@@ -213,13 +176,13 @@ public class TestBindings {
 
         int rowCount = (int) duckdb_vector_size();
         int expected = 0x01020304;
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, rowCount);
+        DuckDBWritableVector writable = new DuckDBWritableVector(vec, rowCount);
         writable.setInt(0, expected);
 
         ByteBuffer rawData = duckdb_vector_get_data(vec, (long) rowCount * Integer.BYTES);
         assertEquals(rawData.order(ByteOrder.nativeOrder()).getInt(0), expected);
 
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(vec, rowCount);
+        DuckDBReadableVector readable = new DuckDBReadableVector(vec, rowCount);
         assertEquals(readable.getInt(0), expected);
 
         duckdb_destroy_vector(vec);
@@ -231,12 +194,12 @@ public class TestBindings {
         ByteBuffer vec = duckdb_create_vector(lt);
 
         int rowCount = (int) duckdb_vector_size();
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, rowCount);
+        DuckDBWritableVector writable = new DuckDBWritableVector(vec, rowCount);
 
         try {
             writable.setInt(0, 42);
             fail("Expected setInt to reject VARCHAR vector");
-        } catch (DuckDBFunctionException exception) {
+        } catch (DuckDBFunctions.FunctionException exception) {
             assertTrue(exception.getMessage().contains("Expected vector type INTEGER, found VARCHAR"));
             assertEquals(exception.getStackTrace()[0].getMethodName(), "setInt");
         }
@@ -263,7 +226,7 @@ public class TestBindings {
             new BigInteger[] {BigInteger.ZERO, new BigInteger("42"), new BigInteger("9223372036854775808"),
                               new BigInteger("18446744073709551615")};
 
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, rowCount);
+        DuckDBWritableVector writable = new DuckDBWritableVector(vec, rowCount);
         for (int i = 0; i < values.length; i++) {
             writable.setUint64(i, values[i]);
         }
@@ -274,7 +237,7 @@ public class TestBindings {
             assertEquals(nativeData.getLong(i * Long.BYTES), values[i].longValue());
         }
 
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(vec, rowCount);
+        DuckDBReadableVector readable = new DuckDBReadableVector(vec, rowCount);
         for (int i = 0; i < values.length; i++) {
             assertEquals(readable.getUint64(i), values[i]);
         }
@@ -293,7 +256,7 @@ public class TestBindings {
             new BigInteger[] {BigInteger.ZERO, new BigInteger("42"), new BigInteger("9223372036854775808"),
                               new BigInteger("340282366920938463463374607431768211455")};
 
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, rowCount);
+        DuckDBWritableVector writable = new DuckDBWritableVector(vec, rowCount);
         for (int i = 0; i < values.length; i++) {
             writable.setUHugeInt(i, values[i]);
         }
@@ -306,7 +269,7 @@ public class TestBindings {
             assertEquals(nativeData.getLong(offset + Long.BYTES), values[i].shiftRight(Long.SIZE).longValue());
         }
 
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(vec, rowCount);
+        DuckDBReadableVector readable = new DuckDBReadableVector(vec, rowCount);
         for (int i = 0; i < values.length; i++) {
             assertEquals(readable.getUHugeInt(i), values[i]);
         }
@@ -423,7 +386,7 @@ public class TestBindings {
         ByteBuffer vec = duckdb_create_vector(lt);
 
         long rowCount = 70;
-        DuckDBWritableVector writable = new DuckDBWritableVectorImpl(vec, rowCount);
+        DuckDBWritableVector writable = new DuckDBWritableVector(vec, rowCount);
         long[] boundaryRows = new long[] {63, 64, 65, rowCount - 1};
         long[] sentinelRows = new long[] {62, 66, 67, 68};
 
@@ -438,7 +401,7 @@ public class TestBindings {
             writable.setNull(row);
         }
 
-        DuckDBReadableVector readable = new DuckDBReadableVectorImpl(vec, rowCount);
+        DuckDBReadableVector readable = new DuckDBReadableVector(vec, rowCount);
         for (long row : boundaryRows) {
             assertTrue(readable.isNull(row));
         }
@@ -451,7 +414,7 @@ public class TestBindings {
             writable.setInt(row, (int) (row + 1000));
         }
 
-        DuckDBReadableVector revalidated = new DuckDBReadableVectorImpl(vec, rowCount);
+        DuckDBReadableVector revalidated = new DuckDBReadableVector(vec, rowCount);
         for (long row : boundaryRows) {
             assertFalse(revalidated.isNull(row));
             assertEquals(revalidated.getInt(row), (int) (row + 1000));
